refactor(popover): popover component is now a functional component
This also refactors the storybook to be functional as well.

perf(popover): makes performance updates to improve render performance

We now debounce the window resize listener to reduce the number of renders when resizing
the window. Additionally, usage of 'getElementPosition' is now memoized, which should minimize
DOM lookups when the component rerenders.